# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Game API",
        "description": "API endpoints for managing games",
        "contact": {
            "email": "example@example.com"
        },
        "license": {
            "name": "MIT License",
            "url": "https://opensource.org/licenses/MIT"
        },
        "version": "1.0.0"
    },
    "paths": {
        "/api/game": {
            "post": {
                "tags": ["Games"],
                "summary": "Create a new game",
                "description": "Creates a new game.",
                "operationId": "7a2506f2777e2d3f3d3e2290e4635687",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": ["user_name", "user_age"],
                                "properties": {
                                    "user_name": {
                                        "type": "string",
                                        "example": "Ultra Test"
                                    },
                                    "user_age": {
                                        "type": "integer",
                                        "example": 30
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Game created successfully",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "game_id": {
                                            "type": "integer"
                                        },
                                        "remaining_time": {
                                            "type": "integer"
                                        },
                                        "api_token": {
                                            "type": "string"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                }
            }
        },
        "/api/game/{gameId}": {
            "put": {
                "tags": ["Games"],
                "summary": "Make a game attempt",
                "description": "Validates and processes a game attempt.",
                "operationId": "fc741c95dd7dc08fdb196339f00b0222",
                "parameters": [
                    {
                        "name": "gameId",
                        "in": "path",
                        "description": "ID of the game",
                        "required": true,
                        "schema": {
                            "type": "integer",
                            "format": "int64"
                        }
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": ["attempt"],
                                "properties": {
                                    "attempt": {
                                        "type": "string"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Game attempt processed successfully",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "attempt": {
                                            "type": "string"
                                        },
                                        "bulls": {
                                            "type": "integer"
                                        },
                                        "cows": {
                                            "type": "integer"
                                        },
                                        "attempts_count": {
                                            "type": "integer"
                                        },
                                        "remaining_time": {
                                            "type": "integer"
                                        },
                                        "evaluation": {
                                            "type": "integer"
                                        },
                                        "ranking": {
                                            "type": "integer"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "400": {
                        "description": "Invalid attempt",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Unauthorized",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Game not found",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "409": {
                        "description": "Duplicate attempt",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "410": {
                        "description": "Game Over",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string"
                                        },
                                        "secret_code": {
                                            "type": "string"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            },
            "delete": {
                "tags": ["Games"],
                "summary": "Delete a game",
                "description": "Delete a game.",
                "operationId": "7a197bc1ac2c5a72b1e9b2ddf6891adc",
                "parameters": [
                    {
                        "name": "gameId",
                        "in": "path",
                        "description": "ID of the game to delete",
                        "required": true,
                        "schema": {
                            "type": "integer",
                            "format": "int64"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Game deleted successfully",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Unauthorized",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Game not found",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                }
            }
        },
        "/api/game/{gameId}/attempts/{attemptNumber}": {
            "get": {
                "tags": ["Games"],
                "summary": "Get attempt response",
                "description": "Retrieves the response of a specific attempt in a game.",
                "operationId": "4ee9e4ce35c9430dd441bb46d6127b1a",
                "parameters": [
                    {
                        "name": "gameId",
                        "in": "path",
                        "description": "ID of the game",
                        "required": true,
                        "schema": {
                            "type": "integer",
                            "format": "int64"
                        }
                    },
                    {
                        "name": "attemptNumber",
                        "in": "path",
                        "description": "Number of the attempt",
                        "required": true,
                        "schema": {
                            "type": "integer",
                            "format": "int64"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Attempt response retrieved successfully",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "attempt": {
                                            "type": "string"
                                        },
                                        "bulls": {
                                            "type": "integer"
                                        },
                                        "cows": {
                                            "type": "integer"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "400": {
                        "description": "Invalid attempt number",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Unauthorized",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Game not found",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                }
            }
        }
    },
    "components": {
        "securitySchemes": {
            "bearerAuth": {
                "type": "http",
                "description": "Authentication with Bearer token",
                "in": "header",
                "name": "Authorization",
                "scheme": "bearer",
                "bearerFormat": "JWT"
            }
        }
    }
}